requirements/requirements.txt
fastapi==0.115.12
python-multipart==0.0.20
asyncio==3.4.3
uvicorn==0.22.0

bcrypt==4.3.0
PyJWT==2.10.1
pymongo==4.13.2
mongoengine==0.29.1
beanie==1.30.0
passlib[bcrypt]==1.7.4

pydantic==2.11.7

python-dotenv==1.1.0
pydantic-settings==2.9.1